Cloud Functions/Tritium/requirements.txt
google-cloud-storage>=1.35.0
google-cloud-secret-manager==2.2.0
sklearn
flask>=1.1.2
pandas>=1.2.1
numpy>=1.20.0
cantools>=36.0.0
joblib>=1.0.0

